+++ sqlbuilder/internal/sel/target.go
@@ -1,0 +1,11 @@
+package sel
+
+type Target interface {
+	SelectTarget() (string, error)
+}
+
+type Table string
+
+func (t Table) SelectTarget() (string, error) {
+	return string(t), nil
+}
